pyproject.toml
[project]
name = "misp-modules"
version = "3.0.1"
description = "MISP modules are autonomous modules that can be used for expansion and other services in MISP"
authors = [
    {name = "Alexandre Dulaunoy", email = "alexandre.dulaunoy@circl.lu"}
]
license = "AGPL-3.0-only"
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
]
requires-python = ">=3.9,<3.13"
dependencies = [
    ## core dependencies
    "orjson",
    "psutil",
    "tornado",
    ## minimum dependencies
    "beautifulsoup4",
    "jinja2",
    "markdownify",
    "pymisp",
    "publicsuffixlist",
    "setuptools",
]

[project.optional-dependencies]
all = [
    ## pinned module dependencies
    "censys==2.0.9",
    "socialscan<2.0.0",
    "yara-python==4.5.0",
    "numpy>=1.26.4,<2.0.0",
    "pandas>=2.0.0",
    "pandas_ods_reader>=1.0.0",
    ## module dependencies
    "apiosintds",
    "assemblyline_client",
    "backscatter",
    "blockchain",
    "clamd",
    "crowdstrike-falconpy",
    "dnsdb2",
    "domaintools_api",
    "geoip2",
    "greynoise",
    "jbxapi",
    "maclookup",
    "matplotlib",
    "mattermostdriver",
    "misp-lib-stix2>=3.0.1.2",
    "misp-stix>=2025.1.10",
    "mwdblib",
    "ndjson",
    "np",
    "oauth2",
    "opencv-python",
    "openpyxl",
    "pandoc",
    "passivetotal",
    "pdftotext",
    "pycountry",
    "pyeti-python3",
    "pyeupi",
    "pygeoip",
    "pyintel471",
    "pyipasnhistory",
    "pymisp[fileobjects, openioc, pdfexport, email]",
    "pypdns",
    "pypssl",
    "pysafebrowsing",
    "pytesseract",
    "python-docx",
    "python-pptx",
    "pyzbar",
    "requests[security]",
    "setuptools",
    "shodan",
    "sigmatools",
    "sigmf",
    "slack-sdk",
    "sparqlwrapper",
    "tau-clients",
    "taxii2-client",
    "trustar",
    "urlarchiver",
    "vt-graph-api",
    "vt-py",
    "vulners",
    "vysion",
    "wand",
    "xlrd",
]

[project.urls]
Issues = "https://github.com/MISP/misp-modules/issues"
Homepage = "https://github.com/MISP/misp-modules"
Documentation = "https://misp.github.io/misp-modules"
Repository = "https://github.com/MISP/misp-modules"

[project.scripts]
misp-modules = "misp_modules.__main__:main"

[tool.poetry]
packages = [{include = "misp_modules"}]

[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
black = "*"
codecov = "*"
flake8 = "*"
flake8-pyproject = "*"
ipdb = "*"
isort = "*"
nose = "*"
pre-commit = "*"
pytest = "*"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "*"
mkdocs-material = "*"
markdown_include = "*"

[tool.poetry.requires-plugins]
poetry-plugin-bundle = ">=1.6"
poetry-plugin-export = ">=1.9"

[tool.black]
line-length = 120
target-version = ["py39"]
preview = true
enable-unstable-feature = ["string_processing"]

[tool.isort]
line_length = 120
profile = "black"

[tool.flake8]
max-line-length = 120
exclude = [".git", "__pycache__"]
extend-ignore = [
    "E402", "E501", "E712", "E203",
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:.*deprecated to return a value that is not None.*:DeprecationWarning:",
]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
